TSDK-188 Implement review by @nandotorterolo
