+++ .gitignore
@@ -1,0 +1,2 @@
+.metals
+scala-daml-broker-app/keyfile.json
