+++ java-daml-lattice-app/README.md
@@ -1,2 +1,2 @@
 # java-daml-lattice-app
-This repository contains a simple application that uses the [daml-dopl-lib](https://github.com/Topl/daml-bifrost-module). 
+This repository contains a simple application that uses the [daml-dopl-lib](https://github.com/Topl/daml-bifrost-module). The lattice application is a simple asset tracking application.
