+++ scala-daml-broker-app/.vscode/launch.json
@@ -1,0 +1,22 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "scala",
+            "request": "launch",
+            "name": "Launch Topl Broker",
+            "mainClass": "broker.BrokerMain",
+            "args": [
+                "127.0.0.1",
+                "6865",
+                "keyfile.json",
+                "test"
+            ],
+            "jvmOptions": [],
+            "env": {}
+        }
+    ]
+}
