+++ scala-daml-broker-app/build.sbt
@@ -15,5 +15,3 @@ lazy val root = (project in file("."))
     libraryDependencies += bramblCommon,
     libraryDependencies += toplDaml
   )
-
-// See https://www.scala-sbt.org/1.x/docs/Using-Sonatype.html for instructions on how to publish to Sonatype.
