--- scala-daml-broker-app/src/test/scala/example/HelloSpec.scala
@@ -1,10 +1,0 @@
-package example
-
-import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should.Matchers
-
-class HelloSpec extends AnyFlatSpec with Matchers {
-  // "The Hello object" should "say hello" in {
-  //   Hello.greeting shouldEqual "hello"
-  // }
-}
